examples/auth/requirements.txt
python-fasthtml
fastcore 
httpx>=0.28.1 
python-dotenv 
pyjwt
cryptography
